requirements.txt
pandas>=2.0.0
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.2.2
xgboost>=1.7.3
torch>=2.0.0
joblib>=1.2.0
streamlit>=1.31.0
pyyaml>=6.0
shap>=0.41.0
pillow>=10.0.0